refact: assets now managed by Webpack.
Instead of each year getting their own asset build approach and their
own set of NPM dependencies, we now have a single package.json, and
Webpack to drive the generation of each year's assets.

This makes managing dependences in the future so much easier, and avoids
a whole tonne of duplication.

All generated files are content-hashed, and so we translate their
locations via the generated public/assets/manifest.json file.